fix broken tests; move some tests from combine to relationship
